+++ src/pages/offer-page/offer-page.tsx
@@ -3,9 +3,9 @@ import { useParams } from 'react-router-dom';
 import { offers } from '../../mocks/offers';
 import { reviews } from '../../mocks/reviews';
 import ReviewForm from '../../components/review-form/review-form';
-import Error404 from '../error404/error404';
-import { Layout } from '../../components/layout/layout';
-import Map from '../../components/map/map';
+import Error404 from '../error/error';
+import { Layout } from '../../components/layout-component/layout-component';
+import Map from '../../components/map-component/map-component';
 import { OfferList } from '../../components/offer-list/offer-list';
 import { Cities } from '../../mocks/city';
 import { OfferGallery } from './offer-gallery/OfferGalery';
